Split topbar view components
