+++ web/src/pages/App/Topbar/IconRenderer.jsx
@@ -1,0 +1,3 @@
+export function IconRenderer({ icon: Icon, size = 18 }) {
+  return <Icon aria-hidden="true" sx={{ fontSize: size }} />;
+}
