+++ web/src/pages/App/Topbar/TopbarLogoLink.jsx
@@ -1,0 +1,29 @@
+import { Box } from "@mui/material";
+
+import { MarkLogo } from "./MarkLogo";
+import { WordmarkLogo } from "./WordmarkLogo";
+
+export function TopbarLogoLink({ ariaLabel, onClick }) {
+  return (
+    <Box
+      component="a"
+      href="/"
+      aria-label={ariaLabel}
+      onClick={onClick}
+      sx={{
+        display: "flex",
+        alignItems: "center",
+        height: 40,
+        flexShrink: 0,
+        textDecoration: "none",
+      }}
+    >
+      <Box sx={{ display: { xs: "block", sm: "none" }, lineHeight: 0 }}>
+        <MarkLogo size={38} framed={false} />
+      </Box>
+      <Box sx={{ display: { xs: "none", sm: "block" } }}>
+        <WordmarkLogo />
+      </Box>
+    </Box>
+  );
+}
