+++ web/src/pages/App/Topbar/index.js
@@ -1,0 +1,1 @@
+export { Topbar } from "./Topbar";
